+++ lib/src/file_system/entries/ftp_file.dart
@@ -1,6 +1,7 @@
 import 'package:pure_ftp/src/extensions/ftp_directory_extensions.dart';
 import 'package:pure_ftp/src/file_system/ftp_entry.dart';
 import 'package:pure_ftp/src/file_system/ftp_file_system.dart';
+import 'package:pure_ftp/src/ftp/exceptions/ftp_exception.dart';
 import 'package:pure_ftp/src/ftp/extensions/ftp_command_extension.dart';
 import 'package:pure_ftp/src/ftp/ftp_commands.dart';
 
@@ -24,11 +25,10 @@ class FtpFile extends FtpEntry {
       if (recursive) {
         await parent.create(recursive: true);
       } else {
-        throw Exception('Parent directory does not exist');
+        throw FtpException('Parent directory does not exist');
       }
     }
-    // TODO: implement create
-    throw UnimplementedError();
+    return _fs.uploadFile(this, []);
   }
 
   @override
@@ -56,16 +56,16 @@ class FtpFile extends FtpEntry {
       return this;
     }
     if (!await newFile.parent.exists()) {
-      throw Exception('Parent directory of new file does not exist');
+      throw FtpException('Parent directory of new file does not exist');
     }
     final response = await FtpCommand.RNFR.writeAndRead(_fs.socket, [path]);
     if (!response.isSuccessful) {
-      throw Exception('Cannot move file');
+      throw FtpException('Cannot move file');
     }
     final response2 =
         await FtpCommand.RNTO.writeAndRead(_fs.socket, [newFile.path]);
     if (!response2.isSuccessful) {
-      throw Exception('Cannot move file');
+      throw FtpException('Cannot move file');
     }
     return newFile;
   }
@@ -73,23 +73,23 @@ class FtpFile extends FtpEntry {
   @override
   Future<FtpFile> rename(String newName) async {
     if (newName.contains('/')) {
-      throw Exception('New name cannot contain path separator');
+      throw FtpException('New name cannot contain path separator');
     }
     final newFile = parent.getChildFile(newName);
     if (newFile.path == path) {
       return this;
     }
     if (!await newFile.parent.exists()) {
-      throw Exception('Parent directory of new file does not exist');
+      throw FtpException('Parent directory of new file does not exist');
     }
     final response = await FtpCommand.RNFR.writeAndRead(_fs.socket, [path]);
     if (!response.isSuccessful) {
-      throw Exception('Cannot rename file');
+      throw FtpException('Cannot rename file');
     }
     final response2 =
         await FtpCommand.RNTO.writeAndRead(_fs.socket, [newFile.path]);
     if (!response2.isSuccessful) {
-      throw Exception('Cannot rename file');
+      throw FtpException('Cannot rename file');
     }
     return newFile;
   }